+++ src/main/java/com/project/moabuja/controller/MemberController.java
@@ -6,30 +6,28 @@ import com.project.moabuja.dto.TokenDto;
 import com.project.moabuja.dto.request.member.MemberUpdateRequestDto;
 import com.project.moabuja.dto.request.member.NicknameValidationRequestDto;
 import com.project.moabuja.dto.response.member.HomeResponseDto;
+import com.project.moabuja.dto.response.member.ReissueDto;
 import com.project.moabuja.security.userdetails.UserDetailsImpl;
-import com.project.moabuja.service.MemberService;
+import com.project.moabuja.service.MemberServiceImpl;
 import com.project.moabuja.util.CustomResponseEntity;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
-import java.nio.charset.Charset;
 
 @RestController
 @RequiredArgsConstructor
 public class MemberController {
 
-    private final MemberService memberService;
+    private final MemberServiceImpl memberServiceImpl;
 
     // 카카오 로그인 api
     @GetMapping("/user/kakao/callback")
     public ResponseEntity kakaoLogin(@RequestParam String code) throws JsonProcessingException {
-        TokenDto dto = memberService.kakaoLogin(code);
+        TokenDto dto = memberServiceImpl.kakaoLogin(code);
             
         // todo : CustomResponseEntity --> authorization 필요없음 , 필요없는거 버리기
         CustomResponseEntity response = CustomResponseEntity.builder()
@@ -47,21 +45,21 @@ public class MemberController {
                                  @AuthenticationPrincipal UserDetailsImpl userDetails) throws JsonProcessingException {
         String email = userDetails.getUsername();
         // 회원이 캐릭터랑 닉네임 설정한 경우
-        return memberService.updateMemberInfo(dto, email);
+        return memberServiceImpl.updateMemberInfo(dto, email);
     }
 
     // 닉네임 이름 중복체크  api
     @PostMapping("/nickname/validation")
     public ResponseEntity nicknameValid(@RequestBody NicknameValidationRequestDto nicknameValidationRequestDto){
-        return memberService.nicknameValid(nicknameValidationRequestDto);
+        return memberServiceImpl.nicknameValid(nicknameValidationRequestDto);
     }
 
     // access 토큰 만료 시 재발급 api : access, refresh 모두 재발급
     @GetMapping("/api/reissue")
     public ResponseEntity reissue(HttpServletRequest request){
-        TokenDto tokenDto = memberService.reissue(request);
+        ReissueDto dto = memberServiceImpl.reissue(request);
         CustomResponseEntity response = CustomResponseEntity.builder()
-                .data(tokenDto)
+                .data(ReissueDto)
                 .message("Redis 저장 성공")
                 .code(HttpStatus.OK)
                 .build();
@@ -70,7 +68,7 @@ public class MemberController {
 
     @GetMapping("/api/logout")
     public ResponseEntity<?> logout(HttpServletRequest request){
-        return memberService.logout(request);
+        return memberServiceImpl.logout(request);
     }
 
 
@@ -78,7 +76,7 @@ public class MemberController {
     public HomeResponseDto getHome(@AuthenticationPrincipal UserDetailsImpl userDetails){
 
         Member currentUser = userDetails.getMember();
-        return memberService.getHomeInfo(currentUser);
+        return memberServiceImpl.getHomeInfo(currentUser);
 
     }
 }
