+++ src/main/java/com/project/moabuja/dto/KakaoUserInfoDto.java
@@ -6,7 +6,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class KakaoUserInfoDto {
-
     private Long kakaoId;
     private String email;
 }
